requirements.txt
sqlalchemy==2.0.20
pydantic==2.8.2
asyncpg==0.28.0
alembic==1.12.0
python-dotenv
fastapi
uvicorn
nanoid